Internalized DeserializationElement into ObjectBuildingVisitor
