+++ jetro-object/src/test/java/net/sf/jetro/object/deserializer/beans/ElementType.java
@@ -1,0 +1,7 @@
+package net.sf.jetro.object.deserializer.beans;
+
+public enum ElementType {
+	ARRAY,
+	OBJECT,
+	PRIMITIVE
+}
